+++ src/MineCase.Protocol/Protocol/Play/EntityHeadLook.cs
@@ -18,7 +18,7 @@ namespace MineCase.Protocol.Play
         public void Serialize(BinaryWriter bw)
         {
             bw.WriteAsVarInt(EID, out _);
-            bw.WriteAsByte(Yaw);
+            bw.WriteAsByte((sbyte)Yaw);
         }
     }
 }
