+++ src/MineCase.Protocol/Protocol/Play/WindowItems.cs
@@ -21,7 +21,7 @@ namespace MineCase.Protocol.Play
 
         public void Serialize(BinaryWriter bw)
         {
-            bw.WriteAsByte(WindowId);
+            bw.WriteAsByte((sbyte)WindowId);
             bw.WriteAsShort(Count);
 
             // FIXME
