+++ src/MineCase.Protocol/Protocol/Play/WindowProperty.cs
@@ -20,7 +20,7 @@ namespace MineCase.Protocol.Play
 
         public void Serialize(BinaryWriter bw)
         {
-            bw.WriteAsByte(WindowId);
+            bw.WriteAsByte((sbyte)WindowId);
             bw.WriteAsShort(Property);
             bw.WriteAsShort(Value);
         }
